+++ crates/amaru/src/bin/amaru/cmd/dev/ledger/mod.rs
@@ -15,9 +15,12 @@
 use clap::Subcommand;
 
 pub(crate) mod convert;
+#[cfg(feature = "mithril")]
+pub(crate) mod mithril;
 pub(crate) mod nonces;
 pub(crate) mod reset;
 pub(crate) mod states;
+pub(crate) mod sync;
 
 #[derive(Debug, Subcommand)]
 pub(crate) enum LedgerCommand {
@@ -34,4 +37,12 @@ pub(crate) enum LedgerCommand {
 
     /// Reset the ledger database to the beginning of a specific epoch.
     Reset(reset::Args),
+
+    /// Sync the ledger with local blocks.
+    ///
+    /// The ledger state must be in an already bootstrapped state (e.g. via Amaru snapshots).
+    Sync(sync::Args),
+
+    #[cfg(feature = "mithril")]
+    Mithril(mithril::Args),
 }